mqtt-bridge/requirements.txt
paho-mqtt==1.6.1
kafka-python==2.0.2 